+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Added
+
+IOC: RegisterGeneric to register open generic types. As(Type) and Keyed(object, Type) now works for open generic types. For now, you need to use \[GenerateFor(typeof(...))]] to manually register all used instances of generic type.
+
 ## 33.3.0
 
 ### Added